feat(dao) reintroduce the quiet option
This commit introduces the 'quiet' flag in the options table
passed to DAO CRUD methods. When set, the quiet flag skips broadcasting
a local 'dao:crud' worker event, which in turn generates a cluster event
regarding the entity, as well as a local cache invalidation of the
entity.

This mimics the 'quiet' functionality present in the legacy DAO layer.